Update Publicize block
